backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
opencv-python-headless>=4.9.0
ultralytics>=8.2.0
mediapipe>=0.10.0
torch>=2.2.0
torchvision>=0.17.0
git+https://github.com/openai/CLIP.git
Pillow>=10.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
pydantic>=2.6.0
python-multipart>=0.0.9
psutil>=5.9.0
slowapi>=0.1.9
pyjwt>=2.8.0
apscheduler>=3.10.0
